+++ .gitignore
@@ -1,12 +1,16 @@
-server/.env
+# Autogenerated directories
 .idea/
 .vscode/
-*.wav
 __pycache__/
+*.egg-info/
+node_modules/
+venv/
 
+# Build directories
 certs/
 dist/
 flagged/
 history/
-node_modules/
-venv/
+
+# Files
+.env
