--- server/enums/features.py
@@ -1,4 +1,0 @@
-from enum import Enum
-class Features(Enum):
-  CHATBOT = 'chat'
-  IMAGE_CAPTIONING = 'caption'
