--- server/enums/logger.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-class LogLevel(Enum):
-  CAPTION = 'CAPTION'
-  CHATBOT = 'CH@'
-  ERROR = 'ERROR'
-  INFO = 'INFO'
-  VOICE_ASSISTANT = 'VAST'
