+++ server/requirements.txt
@@ -1,13 +1,16 @@
 # Python 3.10.15
 
+# Project dependencies
 Flask~=3.0.3
 Flask-Cors~=5.0.0
-gradio~=4.44.0
+gradio~=4.44.1
 numpy~=1.26.4
 pillow~=10.4.0
+pydub~=0.25.1
 python-dotenv~=1.0.1
 requests
 scipy~=1.14.1
+setuptools~=68.2.0
 transformers~=4.38.2
 
 # Model dependencies
@@ -16,3 +19,7 @@ flax
 openai
 tensorflow
 torch
+
+# Project Modules
+-e src/nlp
+-e src/utils
