+++ server/src/nlp/setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='nlp',      # Name of your module
+    version='0.0.1',
+    packages=find_packages(),  # Automatically find all packages
+)
